+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [2.0.0] - 2025-05-01
 - Added user-wide renaming rules:   
 You can now define global renaming rules for a user by placing a `.rename.user.conf` file in the user's root directory. These rules apply to full file paths and are used if no folder-specific `.rename.conf` is present.
 
@@ -33,6 +33,7 @@ You can now define global renaming rules for a user by placing a `.rename.user.c
 ## [1.0.0] - 2025-02-07
 - Initial release.
 
+[2.0.0]: https://github.com/powerflo/files_autorename/releases/tag/v1.7.0
 [1.7.0]: https://github.com/powerflo/files_autorename/releases/tag/v1.7.0
 [1.6.0]: https://github.com/powerflo/files_autorename/releases/tag/v1.6.0
 [1.5.0]: https://github.com/powerflo/files_autorename/releases/tag/v1.5.0